requirements.txt
# Core dependencies
python-dotenv>=1.0.0
google-generativeai>=0.3.0
tweepy>=4.14.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
anyio>=4.2.0